+++ ios/Lift/Device/MultiDevice.swift
@@ -7,10 +7,11 @@ class MultiDevice : DeviceSession, SensorDataDelegate, DeviceDelegate {
     private var deviceInfos: [DeviceId : DeviceInfo] = [:]
     private var deviceInfoDetails: [DeviceId: DeviceInfo.Detail] = [:]
     private var deviceSessions: [DeviceId : DeviceSession] = [:]
-    private var deviceData: [DeviceId : NSMutableData] = [:]
+    private var deviceData: [DeviceId : [NSData]] = [:]
     private var devices: [ConnectedDevice] = []
     private let combinedStats = DeviceSessionStats<DeviceSessionStatsTypes.KeyWithLocation>()
-    private let deviceId = NSUUID(UUIDString: "00000000-0000-0000-0000-000000000000")!
+    private let deviceId = DeviceId(UUIDString: "00000000-0000-0000-0000-000000000000")!
+    private var zeroed: Bool = false
     
     private var sensorDataDelegate: SensorDataDelegate!
     private var deviceDelegate: DeviceDelegate!
@@ -103,11 +104,32 @@ class MultiDevice : DeviceSession, SensorDataDelegate, DeviceDelegate {
     
     // MARK: SensorDataDelegate implementation
     
-    func sensorDataEnded(deviceId: NSUUID, deviceSession: DeviceSession) {
+    func sensorDataNotReceived(deviceId: DeviceId, deviceSession: DeviceSession) {
+        // never called
+    }
+    
+    func sensorDataEnded(deviceId: DeviceId, deviceSession: DeviceSession) {
         sensorDataDelegate.sensorDataEnded(deviceId, deviceSession: self)
     }
     
-    func sensorDataReceived(deviceId: NSUUID, deviceSession: DeviceSession, data: NSData) {
+    func sensorDataReceived(deviceId: DeviceId, deviceSession: DeviceSession, data: NSData) {
+        // when a packet arrives except from this device, we will zero all devices to this time
+        // we say except for this, because it is more important to zero the devices communicating
+        // with this device externally. This device can always be reset immediately; and if it 
+        // is the only device we have, then it does not need zeroing anyway.
+        if deviceId != ThisDevice.Info.id {
+            if !zeroed {
+                for d in devices { d.zero() }
+                zeroed = true
+            }
+        }
+        
+        if let ds = deviceData[deviceId] {
+            deviceData[deviceId] = ds + [data]
+        } else {
+            deviceData[deviceId] = [data]
+        }
+        
         combinedStats.merge(deviceSession.getStats()) { k in
             let location = k.deviceId == ThisDevice.Info.id ? DeviceInfo.Location.Waist : DeviceInfo.Location.Wrist
             return DeviceSessionStatsTypes.KeyWithLocation(sensorKind: k.sensorKind, deviceId: k.deviceId, location: location)
