+++ ios/Lift/Device/Pebble/PebbleDevice.swift
@@ -1,54 +1,5 @@
 import Foundation
 
-/**
- * Session receiving app updates from the watch. When initialised with ``PBWatch``, it sets all required handlers
- * and receives the incoming data from the device.
- *
- * It also maintains required stats.
- */
-class PebbleDeviceSession : DeviceSession {
-    private var sensorDataDelegate: SensorDataDelegate!
-    private var startTime: NSDate!
-    private var updateHandler: AnyObject?
-    private var watch: PBWatch!
-    private var deviceId: NSUUID!
-
-    required init(deviceId: NSUUID, watch: PBWatch!, sensorDataDelegate: SensorDataDelegate) {
-        super.init()
-        self.watch = watch
-        self.deviceId = deviceId
-        self.startTime = NSDate()
-        self.sensorDataDelegate = sensorDataDelegate
-        self.updateHandler = watch.appMessagesAddReceiveUpdateHandler(appMessagesReceiveUpdateHandler)
-    }
-    
-    // MARK: main
-    
-    override func stop() {
-        if let x: AnyObject = updateHandler { watch.appMessagesRemoveUpdateHandler(x) }
-        sensorDataDelegate.sensorDataEnded(deviceId, deviceSession: self)
-    }
-    
-    private func appMessagesReceiveUpdateHandler(watch: PBWatch!, data: [NSObject : AnyObject]!) -> Bool {
-        let adKey = NSNumber(uint32: 0xface0fb0)
-        let deadKey = NSNumber(uint32: 0x0000dead)
-        if let x = data[adKey] as? NSData {
-            accelerometerDataReceived(x)
-        } else if let x: AnyObject = data[deadKey] {
-            stop()
-        }
-        return true
-    }
-    
-    private func accelerometerDataReceived(data: NSData) {
-        updateStats(DeviceSessionStatsTypes.Key(sensorKind: .Accelerometer, deviceId: deviceId), update: { prev in
-            return DeviceSessionStatsTypes.Entry(bytes: prev.bytes + data.length, packets: prev.packets + 1)
-        })
-        
-        sensorDataDelegate.sensorDataReceived(deviceId, deviceSession: self, data: data)
-    }
-}
-
 class PebbleDevice : NSObject, Device {
     internal let central = PBPebbleCentral.defaultCentral()
     internal let pebbleDeviceType = "pebble"
@@ -142,6 +93,13 @@ class PebbleConnectedDevice : PebbleDevice, PBPebbleCentralDelegate, PBWatchDele
     
     func stop() {
         findWatch().either({ x in self.deviceDelegate.deviceDidNotConnect(x) }, onR: { $0.appMessagesKill(self.appKilled) })
+    }
+    
+    func zero() -> NSTimeInterval {
+        if let x = currentDeviceSession {
+            x.zero()
+        }
+        return 0.1  // on average, we take 100 milliseconds
     }
     
     // MARK: PBPebbleCentralDelegate implementation
