+++ ios/Lift/Lift-Bridging-Header.h
@@ -2,6 +2,8 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
+#import "Device/This/ThisDeviceProtocol.h"
+
 #import <PebbleKit/PebbleKit.h>
 #import <CommonCrypto/CommonDigest.h>
 #import <AudioToolbox/AudioToolbox.h>
